requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.6
pandas==2.2.0
openpyxl==3.1.2
langchain==0.1.5
langchain-openai==0.0.5
python-dotenv==1.0.1 